requirements.txt
Django
Pillow
bleach
gunicorn
sendgrid
requests
psycopg2
whitenoise
python-dotenv
dj-database-url
django-cors-headers
djangorestframework